refactor: move types out to admin root
