+++ packages/strapi-plugin-rich-text/admin/src/types.ts
